requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-login
pandas
numpy
scikit-learn
python-dotenv
gunicorn
psycopg2-binary
